// tools/image-palette-extractor/client/types.ts
export type ImagePaletteExtractorMessages = Readonly<{
  meta: {
    name: string
    description: string
  }
  uploadTitle: string
  uploadDescription: string
  chooseImageLabel: string
  uploadHint: string
  supportedFormatsLabel: string
  changeImageLabel: string
  removeImageLabel: string
  optionsTitle: string
  optionsDescription: string
  colorCountLabel: string
  colorCountDescription: string
  qualityLabel: string
  qualityDescription: string
  qualityFast: string
  qualityBalanced: string
  qualityPrecise: string
  sortLabel: string
  sortDescription: string
  sortDominance: string
  sortHue: string
  sortLightness: string
  ignoreTransparentLabel: string
  ignoreTransparentDescription: string
  resultsTitle: string
  resultsDescription: string
  emptyResultTitle: string
  emptyResultDescription: string
  loadingTitle: string
  loadingDescription: string
  extractedColorsLabel: string
  sampledPixelsLabel: string
  dominantColorLabel: string
  exportTitle: string
  exportDescription: string
  exportFormatLabel: string
  exportFormatHex: string
  exportFormatCss: string
  exportFormatJson: string
  copyPaletteLabel: string
  copiedPaletteLabel: string
  downloadPaletteLabel: string
  copyColorLabel: string
  copiedColorLabel: string
  errorTitle: string
  invalidFileTypeError: string
  loadFailedError: string
  noPixelsError: string
}>